requirements.txt
Django==1.9.8
South==1.0.2
argparse==1.2.1
dj-database-url==0.4.1
django-allauth==0.27.0
django-bootstrap3==7.1.0
django-bower==5.1.0
django-email-registration==0.3.1
django-filter==0.15.0
future==0.15.2
gunicorn==19.6.0
oauthlib==2.0.0
psycopg2==2.6.2
pycrypto==2.6.1
pycryptodome==3.4
pyjwkest==0.6.2
python-openid==2.2.5
requests==2.11.1
requests-oauthlib==0.7.0
six==1.10.0
towel==0.8.1
whitenoise==3.2
wsgiref==0.1.2
